Debug get_read1_ai_action

+++ neural_network/nn_shogi.c
@@ -166,10 +166,10 @@ Action get_read1_ai_action(NNAI *self, const Game *game) {
     int len_all_actions = get_useful_actions_with_tfr(game, all_actions);
 
     int best_action = 0;
-    int min_evaluation = 1.0;
+    double min_evaluation = 1.0;
     for (int i = 0; i < len_all_actions; i++) {
         do_action((Game *) game, all_actions[i]);
-        int evaluation = nn_evaluate(&self->nn, 1-game->turn%2, &game->current);
+        double evaluation = nn_evaluate(&self->nn, 1-game->turn%2, &game->current);
         undo_action((Game *) game);
         if (evaluation < min_evaluation) {
             best_action = i;
